+++ tests/manifest/fragments/e2e.ts
@@ -1,0 +1,3 @@
+import { TestEntry } from '../test-types';
+
+export const e2eTests: TestEntry[] = [];