+++ tests/manifest/fragments/unit.ts
@@ -1,0 +1,3 @@
+import { TestEntry } from '../test-types';
+
+export const unitTests: TestEntry[] = [];