+++ tests/manifest/fragments/vendor-device.ts
@@ -1,0 +1,3 @@
+import { TestEntry } from '../test-types';
+
+export const vendor_deviceTests: TestEntry[] = [];